更いただいた点をを変更更しまし
