+++ apps/user-api/.env.example
@@ -7,7 +7,7 @@ DB_DIALECT="mysql"
 DB_DRIVER="pymysql"
 DB_HOST="db"
 DB_PORT="3306"
-DB_DATABASE="sqlqlchemy"
+DB_DATABASE="sqlalchemy"
 DB_USERNAME="root"
 DB_PASSWORD="secret"
 DB_CHARSET_TYPE="utf8"
