+++ apps/user-api/app/models/order_details.py
@@ -2,5 +2,5 @@ from common.database.model_mixins.order_mixins import OrderDetailMixin
 from config.settings import Base
 
 
-class Order(Base, OrderDetailMixin):
+class OrderDetail(Base, OrderDetailMixin):
     __tablename__ = "order_details"
