+++ apps/user-api/app/models/products.py
@@ -1,7 +1,7 @@
-from common.database.model_mixins.order_mixins import ProductMixin
+from common.database.model_mixins.product_mixins import ProductMixin
 from config.settings import Base
 
 
-class Order(Base, ProductMixin):
+class Product(Base, ProductMixin):
     __tablename__ = "products"
