+++ common/database/model_mixins/__init__.py
@@ -1,3 +1,4 @@
-from .user_mixins import UserMixin
-from .product_mixins import ProductMixin
-from .order_mixins import OrderMixin
+from .user_mixins import *
+from .product_mixins import *
+from .order_mixins import *
+
