+++ common/database/model_mixins/order_mixins.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 from pytz import timezone  # type: ignore
-from sqlalchemy import BigInteger, Column, DateTime, String, Integer, ForeginKey
+from sqlalchemy import BigInteger, Column, DateTime, String, Integer, ForeignKey
 from sqlalchemy.orm import declarative_mixin
 
 from config.settings import TIME_ZONE
@@ -14,7 +14,7 @@ def current_timestamp():
 class OrderMixin:
     id = Column(BigInteger, primary_key=True)
     code = Column(String(255), unique=True, nullable=False)
-    user_id = Column(BigInteger, ForeginKey("users.id"),nullable = False)
+    user_id = Column(BigInteger, ForeignKey("users.id"),nullable = False)
     post_code = Column(String(10), nullable=False)
     user_address1 = Column(String(255), nullable=False)
     user_address2 = Column(String(255), nullable=False)
@@ -35,3 +35,21 @@ class OrderMixin:
     updated_at = Column(
         DateTime, nullable=False, default=current_timestamp, onupdate=current_timestamp
         )
+
+
+@declarative_mixin
+class OrderDetailMixin:
+    id = Column(BigInteger, primary_key=True)
+    order_id = Column(BigInteger, ForeignKey("orders.id"), nullable=False)
+    product_id = Column(BigInteger, ForeignKey("products.id"), nullable=False)
+    product_code = Column(String(255), nullable=False)
+    product_name = Column(String(255), nullable=False)
+    product_unit_price = Column(BigInteger, nullable=False)
+    product_percent_tax = Column(String(255), nullable=False)
+    purchase_number = Column(BigInteger, nullable=False)
+    price = Column(BigInteger, nullable=False)
+    created_at = Column(DateTime, nullable=False, default=current_timestamp)
+    updated_at = Column(
+        DateTime, nullable=False, default=current_timestamp, onupdate=current_timestamp
+        )
+
