+++ common/database/model_mixins/product_mixins.py
@@ -16,7 +16,7 @@ def current_timestamp():
 class ProductMixin:
     id = Column(BigInteger, primary_key=True)
     code = Column(String(255), nullable=False, unique=True, comment="商品コード")
-    name = Column(String(255), comment="商品名") # name って nullable=Falseにしなくていいのか気になります
+    name = Column(String(255), nullable=False, comment="商品名")
     unit_price = Column(BigInteger, nullable=False, comment="単価")
     percent_tax = Column(String(255), nullable=False, comment="消費税率")
     created_at = Column(DateTime, nullable=False, default=current_timestamp)
