+++ docker/setup.dev.sql
@@ -2,7 +2,7 @@ CREATE USER 'root' @'%' IDENTIFIED WITH mysql_native_password BY 'secret';
 
 GRANT ALL PRIVILEGES ON *.* to 'root' @'%';
 
-CREATE DATABASE `sqlqlchemy` DEFAULT CHARACTER SET utf8mb4;
+CREATE DATABASE `sqlalchemy` DEFAULT CHARACTER SET utf8mb4;
 
 CREATE DATABASE `test` DEFAULT CHARACTER SET utf8mb4;
